requirements.txt
python-telegram-bot
python-dotenv
ollama
nest_asyncio